fix(plugin-core): delete same type
